requirements.txt
flask
yt-dlp
ffmpeg-python
nltk
scikit-learn
transformers
vosk
python-dotenv